requirements.txt
rumps
requests
pyobjc
py2app
